code from 1/22/2020
 add shooter
 add indexer

+++ src/main/java/frc/robot/commands/shooter/ShooterVelocity.java
@@ -7,9 +7,9 @@ import frc.robot.subsystems.Shooter;
 public class ShooterVelocity extends CommandBase {
   private Shooter subsystem = RobotContainer.SHOOTER;
 
-  private final int output;
+  private final double output;
 
-  public ShooterVelocity(int velocity) {
+  public ShooterVelocity(double velocity) {
     addRequirements(subsystem);
     this.output = velocity;
   }
